requirements.txt
openai-whisper==20231117
yt-dlp==2024.3.10
torch
torchaudio
numpy
tqdm
ffmpeg-python